[minor] query fixes
