+++ erpnext/patches/v8_3/update_company_total_sales.py
@@ -3,11 +3,13 @@
 
 from __future__ import unicode_literals
 import frappe
-from erpnext.accounts.doctype.sales_invoice.sales_invoice import update_company_monthly_sales
+from erpnext.setup.doctype.company.company import update_company_current_month_sales, update_company_monthly_sales
 
 def execute():
-	'''Update company monthly sales field based on sales invoices this month'''
+	'''Update company monthly sales history based on sales invoices'''
+	frappe.reload_doctype("Company")
 	companies = [d['name'] for d in frappe.get_list("Company")]
 
 	for company in companies:
+		update_company_current_month_sales(company)
 		update_company_monthly_sales(company)
